% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_knn_cv.R
\name{my_knn_cv}
\alias{my_knn_cv}
\title{K-nearest Neighbors function}
\usage{
my_knn_cv(train, cl, k_nn, k_cv)
}
\arguments{
\item{train}{the training data set}

\item{cl}{factor of true classifications of training set}

\item{k_nn}{number of neighbours considered}

\item{k_cv}{integer representing the number of folds}
}
\value{
a list with a vector of the predicted class and
  a numeric with the cross-validation misclassification rate
  a numeric with the training misclassification rate
}
\description{
This function uses k-nearest neighbour classification for test set from training set.
}
\examples{
my_knn_cv(my_iris[, -5], my_iris$Species, 1, 5)

}
\keyword{prediction}
