% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_lm.R
\name{my_lm}
\alias{my_lm}
\title{Fitting Linear Model function}
\usage{
my_lm(fnc, data)
}
\arguments{
\item{x}{Numeric input to be raised to the power of \code{power}}

\item{power}{Numeric input for the power that \code{x} will be raised to,
default to \code{2}}
}
\value{
Numeric represrnting \code{x} raised to the power of \code{power}
}
\description{
This function raises input to a power
}
\examples{
my_pow(4)
my_pow(4,power = 3)

}
