% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_t_test.R
\name{my_t.test}
\alias{my_t.test}
\title{Test function}
\usage{
my_t.test(x, alternative, mu)
}
\arguments{
\item{x}{Numeric vector input of data.}

\item{alternative}{A character string specifying the alternative hypothesis,
must be one of \code{two.sided}, \code{greater} or \code{less}.}

\item{mu}{A number indication the null hupothesis value of mean.}
}
\value{
A list contains the numeric test statistic, the degree of freedom,
  the value of the parameter \code{alternative}, and the numeric \code{p-value}.
}
\description{
This function performs one and two sample t-tests on vectors of data.
}
\examples{
my_t.test(my_iris$Sepal.Length, "less", 5)

}
