% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_knn_cv.R
\name{my_knn_cv}
\alias{my_knn_cv}
\title{K-nearest Neighbors function}
\usage{
my_knn_cv(train, cl, k_nn, k_cv)
}
\arguments{
\item{x}{Numeric input to be raised to the power of \code{power}}

\item{power}{Numeric input for the power that \code{x} will be raised to,
default to \code{2}}
}
\value{
Numeric represrnting \code{x} raised to the power of \code{power}
}
\description{
This function raises input to a power
}
\examples{
my_knn_cv(my_iris[, -5], my_iris$Species, 1, 5)

}
