% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_lm.R
\name{my_lm}
\alias{my_lm}
\title{Fitting Linear Model function}
\usage{
my_lm(fnc, data)
}
\arguments{
\item{fnc}{An object of class "formula": a symbolic description
of the model to be fitted.}

\item{data}{An optional data frame, list or environment
(or object coercible by as.data.frame to a data frame) containing
the variables in the model.}
}
\value{
Numeric represrnting the coefficients and statistics of the fitting
}
\description{
This function is used to fit linear models. It can be
used to carry out regression, single stratum analysis
of variance and analysis of covariance.
}
\examples{
my_model <- my_lm(Sepal.Length~Sepal.Width, data = my_iris)

}
\keyword{inference}
