% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_rf_cv.R
\name{my_rf_cv}
\alias{my_rf_cv}
\title{Random Forest function}
\usage{
my_rf_cv(k)
}
\arguments{
\item{x}{Numeric input to be raised to the power of \code{power}}

\item{power}{Numeric input for the power that \code{x} will be raised to,
default to \code{2}}
}
\value{
Numeric represrnting \code{x} raised to the power of \code{power}
}
\description{
This function raises input to a power
}
\examples{
my_rf_cv(5)

}
